Refactored how components are stored and processed.

+++ Platformer Test/Tile.cpp
@@ -1,6 +1,6 @@
 #include "Tile.h"-
 
-Tile::Tile(sf::Texture& tex, sf::Rect<int>& rect, TileType type, int x, int y) :
+Tile::Tile(sf::Texture& tex, sf::Rect<int>& rect, TileType type, float x, float y) :
 	m_texture(tex), m_rect(rect), m_tileType(type), m_xPos(x), m_yPos(y) {
 
 	m_sprite.setTexture(m_texture);
